Added footer to objects page
