+++ frontend-shared/src/visualizer-app/knowledge-base/main.styl
@@ -22,6 +22,8 @@ body, html
   min-height 100vh
   display flex
   flex-direction: column
+  div.spacer
+    flex-grow 1
   &>.results
     flex-grow 1
     display flex
